// src/screens/UserFeedScreen/UserFeedScreen.types.ts
import { UserWorkout, Workout, Exercise } from 'src/types/api-autogenerated';

export type NormalizedWorkout = {
  items: string[];
  data: Record<
    string,
    UserWorkout & {
      workout: Workout & { exercises: Exercise[] };
    }
  >;
};
export type WeeklyProgress = {
  target: number; // target workouts per week
  completed: number; // completed workouts this week
  streak: number; // current workout streak
  calories: number; // calories burned this week
};